Add tomato size option
In all good US converters there must be a way to provide
an answer in Tomatoes - this is afterall standard form when converting miles to
feet.
I resisted the tempation to special case 528m.

The size here comes from Wikipedia (sorry) - "The size of
"the tomato varies according to the cultivar, with a range of
1–10 cm (1⁄2–4 in) in width."